+++ src/V1/refinement/RefinedNumber.ts
@@ -41,10 +41,10 @@ import { RefinedPrimitive } from "./RefinedPrimitive";
  * The DataGuarantee and OnError handler are passed into the base class's
  * constructor().
  *
- * `EX` is the DataGuarantee's onError EX type
+ * `EX` is the type of extra data passed to the OnError handler
  */
-export class RefinedNumber<E = object> extends RefinedPrimitive<number, E> {
-    public [Symbol.toPrimitive](hint: string) {
+export class RefinedNumber<EX = object> extends RefinedPrimitive<number, EX> {
+    public [Symbol.toPrimitive](hint: string): string|number {
         if (hint === "string") {
             return this.value.toString();
         }