+++ src/V1/refinement/RefinedString.ts
@@ -41,10 +41,10 @@ import { RefinedPrimitive } from "./RefinedPrimitive";
  * The DataGuarantee and OnError handler are passed into the base class's
  * constructor().
  *
- * `EX` is the DataGuarantee's onError EX type
+ * `EX` is the type of extra data passed to the OnError handler
  */
-export class RefinedString<EX> extends RefinedPrimitive<string, EX> {
-    public [Symbol.toPrimitive](hint: string) {
+export class RefinedString<EX = object> extends RefinedPrimitive<string, EX> {
+    public [Symbol.toPrimitive](hint: string): string|null {
         if (hint === "number") {
             return null;
         }